refactoring structs for better maintainance
